Lint/sonar tidy (#65)
* refactor: reduce complexity and modernize syntax across codebase

Simplify using directives, apply target-typed new expressions,
remove unnecessary fully-qualified type names, and clean up
redundant code patterns.

* refactor: simplify IsEmpty with LINQ, discard unused lambda param, bump Polyfill

* docs: architecture

+++ XLibur/Extensions/XmlWriterExtensions.cs
@@ -39,7 +39,7 @@ internal static class XmlWriterExtensions
             // Cast to long to avoid boxing via XmlWriter.WriteValue(object).
             // XmlWriter has a WriteValue(long) overload but not WriteValue(uint).
             w.WriteStartAttribute(attrName);
-            w.WriteValue((long)value);
+            w.WriteValue(value);
             w.WriteEndAttribute();
         }
